AP-V1

+++ airportsADT.h
@@ -1,16 +1,27 @@
 #ifndef _POSCDT_H_
+
 #define _POSCDT_H_
+
 #include <stdlib.h>
 
+
+
 typedef struct flightCDT * flightADT;
+
 typedef struct airportCDT * airportADT;
+
 typedef struct airlineQueryCDT * airlineQueryADT;
+
 typedef struct destinationCDT * destinationADT;
+
 typedef struct internationalCDT * internationalADT;
 
 typedef struct airline * tAirline;
+
 typedef struct airport * tAirport;
+
 typedef struct flight * tFlight;
+
 typedef struct destination * tDestination;
 
 enum days {SUN=0, MON, TUE, WED, THU, FRI, SAT};
@@ -25,19 +36,20 @@ struct flight
     char * orOaci;		/* No puede estar en blanco */
     char * dstOaci;		/* No puede estar en blanco y puede contener codigos que no sean aeropuertos conocidos */
     char * airline;
+
 };
 
 struct airport
 {
-	char * oaci;
+    char * oaci;
 	char local[4];
 	char * name;
 	char iata[4];
 	int type;
 	int condition;
 	int isInternational;
-	size_t takeOffs;
-	size_t landings;
+	unsigned int takeOffs;
+	unsigned int landings;
 	enum days day;
 	destinationADT destinations;	/* Lista de struct destination */
 };
@@ -45,14 +57,13 @@ struct airport
 struct destination
 {
 	char * oaci; /* debo poner char * por los codigos del tipo AR-XXXX sino seria char dest[5]*/
-	size_t landings;
-	size_t takeOffs;
+	unsigned int landings;
+	unsigned int takeOffs;
 };
 
 struct airline
 {
 	char * name;
-	size_t movs;
+	unsigned int movs;
 };
-
 #endif
